General Cleaning
Added templating
Added essential functions for templating
Added temporary .htaccess
