+++ .htaccess
@@ -1,0 +1,17 @@
+RewriteEngine On
+RewriteBase /wbcts/
+
+
+# Redirect all requests from /subdirectory to root
+RewriteCond %{REQUEST_URI} ^/views(/.*)?$ [NC]
+RewriteRule ^(.*)$ /$1 [R=301,L]
+
+# Rewrite all requests to subdirectory folder
+RewriteCond %{REQUEST_FILENAME} !-f
+RewriteCond %{REQUEST_FILENAME} !-d
+RewriteRule ^(.*)$ views/$1 [L]
+
+RewriteCond %{REQUEST_FILENAME} !-f
+RewriteCond %{REQUEST_FILENAME} !-d
+RewriteRule ^(.*)$ views/$1 [L]
+RewriteRule ^$ views/index.php [L]
