+++ php/global.php
@@ -1,4 +1,5 @@
 <?php
+$WEB_APP_ROOT = "wirm/";
 require_once "php/config/Database.php";
 
 $database = new Database();
