--- profile.php
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<html>
-
-<head>
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<title>Blotter Website | Profile</title>
-
-	<link rel="stylesheet" href="css/style.css">
-</head>
-
-
-<body>
-	<?php
-    include("header.php");
-	?>
-
-
-
-
-	<?php
-    include("footer.php");
-	?>
-</body>
-
-</html>
