--- qr-code.php
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<html>
-
-<head>
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<title>Blotter Website | QR CODE</title>
-
-	<link rel="stylesheet" href="css/style.css">
-</head>
-
-
-<body>
-	<?php
-    include("header.php");
-	?>
-
-	<br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br>
-	<center><p>
-		This is for the QR CODE TRACKER
-	</p></center>
-
-
-
-	<?php
-    include("footer.php");
-	?>
-</body>
-
-</html>
