+++ views/components/footer.php
@@ -1,0 +1,17 @@
+<footer class="footer-distributed">
+	<div class="footer-right">
+		<a href="#"><i class="fab fa-facebook"></i></a>
+		<a href="#"><i class="fab fa-twitter"></i></a>
+		<a href="#"><i class="fab fa-instagram"></i></a>
+	</div>
+	<div class="footer-left">
+		<p class="footer-links">
+			<a class="link-1" href="index.php">Home</a>
+			<a href="OnSettlements.php">Settlements</a>
+			<a href="blotter-records.php">Blotter Records</a>
+			<a href="qr-code.php">QR Code Tracking</a>
+			<a href="profile.php">Profile</a>
+		</p>
+		<p>Blotter Website 2023</p>
+	</div>
+</footer>
