+++ views/qr-code.php
@@ -1,0 +1,16 @@
+<?php
+require("../templates/template-functions.php");
+
+
+$title = "QR Tracking";
+$content = <<<CONTENT
+	
+	<br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br>
+	<center>
+		<p>
+			This is for the QR CODE TRACKER
+		</p>
+	</center>
+CONTENT;
+?>
+<?php include '../templates/default.php'; ?>
